Move search related logic to custom useSearch hook
